requirements.txt
pandas>=1.5.0
requests>=2.28.0
urllib3>=1.26.0
python-dotenv>=0.19.0
numpy>=1.21.0 